Reorder let bindings.
This is done so that the code generated by Camlp4 in OCaml 3.11 works as
expected. The following revised syntax code is parsed differently depending
on the OCaml version:

  <:expr<
    do {
      let a = "foo" in
      print_endline a;
      print_endline a
    }
  >>

In OCaml 3.10, this is translated to

  let a = "foo" in (print_endline a; print_endline a)

In OCaml 3.11, it becomes

  ((let a = "foo" in print_endline a); print_endline a)

in which `a' is out of of scope in the second print_endline statement.
We move the let binding out of the do block so that in both versions the
code becomes

  let a = "foo" in (print_endline a; print_endline a)